requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
marshmallow==3.17.0
    # via -r requirements.in
packaging==21.3
    # via marshmallow
pyparsing==3.0.9
    # via packaging
speedtest-cli==2.1.3
    # via -r requirements.in
